+++ src/main/kotlin/io/openfuture/chain/entity/transaction/BaseTransaction.kt
@@ -14,7 +14,7 @@ abstract class BaseTransaction(
     var timestamp: Long,
 
     @Column(name = "amount", nullable = false)
-    var amount: Double,
+    var amount: Long,
 
     @Column(name = "recipient_address", nullable = false)
     var recipientAddress: String,
